requirements.txt
streamlit
pytesseract
Pillow
numpy
sounddevice
TTS
